requirements.txt
# AITI Assistant - Dependencies
# Python 3.11+

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# LLM & Embeddings
openai==1.12.0
anthropic==0.18.1
tiktoken==0.6.0

# Vector Store & Database
chromadb==0.4.22
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
unstructured==0.12.4

# Telegram Bot
python-telegram-bot==21.0.1

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.2.1
httpx==0.26.0
aiofiles==23.2.1

# Logging & Monitoring
structlog==24.1.0
prometheus-client==0.19.0

# Development
pytest==8.0.0
pytest-asyncio==0.23.4
black==24.2.0
ruff==0.2.1
